Moving the random location generator
